DroidLogicTvInput: add new activity to search channel [6/6]
PD# 149733

add new activity to search channel

+++ DroidLogicTvInput/src/com/droidlogic/tvinput/settings/SettingsManager.java
@@ -197,7 +197,7 @@ public class SettingsManager {
 
         if (mTvSource == TvControlManager.SourceInput_Type.SOURCE_TYPE_TV
             || mTvSource == TvControlManager.SourceInput_Type.SOURCE_TYPE_DTV) {
-            long channelId = intent.getIntExtra(DroidLogicTvUtils.EXTRA_CHANNEL_NUMBER, -1);
+            long channelId = intent.getLongExtra(DroidLogicTvUtils.EXTRA_CHANNEL_NUMBER, -1);
             currentChannel = mTvDataBaseManager.getChannelInfo(TvContract.buildChannelUri(channelId));
             if (currentChannel != null) {
                 Log.d(TAG, "current channel is: ");
